# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Veil Panel API
  version: "0.6.3"
  description: |
    HTTP management API exposed by the Veil Panel (`veil serve`).

    ## Authentication model
    Veil is fail-closed for public Panel exposure. A non-loopback `--listen`
    address requires native TLS, API token authentication (`--auth-token` or
    `VEIL_API_TOKEN`), user/session authentication, and authenticated metrics.
    Caddy exposure requires user/session authentication and authenticated
    metrics even though the Veil listener remains on loopback.

    An unconfigured local loopback listener exposes only the first-run setup
    flow needed to create the initial administrator. Setup is unavailable
    through direct or Caddy exposure.

    `/api/*` routes accept either `X-Veil-Token`, `Authorization: Bearer`, or
    a `veil_session` cookie created by `/api/auth/login`. Mutating requests
    made with a cookie session must also send `X-CSRF-Token`, and only `admin`
    sessions may mutate management state. The sole viewer mutation exception
    is `POST /api/auth/locale`, which updates only the authenticated user's own
    language preference and rejects static API-token callers.

    Browser sessions survive Panel restarts. Only hashes of the session cookie
    and CSRF token are persisted. Sessions expire after 30 minutes idle or 24
    hours absolute and are revoked when a user's password, role, or existence
    changes.

...
    until a valid session exists.

...
  license:
...
servers:
...
security:
  - veilToken: []
  - bearerAuth: []
  - sessionCookie: []
...
paths:
...
  /api/backups:
    get:
      x-veil-role: admin-metadata
      tags: [backups]
      summary: List managed disaster-recovery archives
      description: Requires an admin token or admin session.
      responses:
        "200":
          description: Archives in newest-first order.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BackupArchive"
              examples:
                managed:
                  value:
                    - name: veil_backup_20260605_020000.tar.gz.enc
                      path: /var/lib/veil/backups/veil_backup_20260605_020000.tar.gz.enc
                      size: 4096
                      createdAt: "2026-06-05T02:00:00Z"
                      encrypted: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      x-veil-role: admin-mutation
      tags: [backups]
      summary: Create and verify an encrypted archive
      description: |
        Uses `/etc/veil/backup.passphrase` server-side. The passphrase is never
        accepted from or returned to the browser. Requires admin and CSRF for a
        cookie session.
      security:
        - veilToken: []
        - bearerAuth: []
        - sessionCookie: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BackupCreateRequest"
            examples:
              create-and-prune:
                value:
                  prune: true
                  daily: 7
                  weekly: 4
                  monthly: 12
      responses:
        "201":
          description: Archive created, persisted, and verified.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BackupCreateResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
components:
  securitySchemes:
    veilToken:
      type: apiKey
      in: header
      name: X-Veil-Token
    bearerAuth:
      type: http
      scheme: bearer
    sessionCookie:
      type: apiKey
      in: cookie
      name: veil_session
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
...
  responses:
    BadRequest:
      description: Request body, query string, or path failed validation.
      content:
        text/plain:
          schema:
            $ref: "#/components/schemas/ErrorText"
          examples:
            invalid:
              value: "invalid JSON\n"
    Unauthorized:
      description: Missing or invalid token/session.
      headers:
        WWW-Authenticate:
          schema:
            type: string
          description: Present on authenticated API challenges.
      content:
        text/plain:
          schema:
            $ref: "#/components/schemas/ErrorText"
          examples:
            missing:
              value: "missing or invalid API token or session\n"
    Forbidden:
      description: Authenticated principal is not allowed to perform the operation.
      content:
        text/plain:
          schema:
            $ref: "#/components/schemas/ErrorText"
          examples:
            csrf:
              value: "invalid or missing CSRF token\n"
...
    ServiceUnavailable:
      description: Runtime dependency is unavailable.
      content:
        text/plain:
          schema:
            $ref: "#/components/schemas/ErrorText"
...
  schemas:
...
    ErrorText:
      type: string
      description: Plain text error emitted by `writeError`; usually a single message line ending in `\n`.
...
    BackupArchive:
      type: object
      required: [name, path, size, createdAt, encrypted]
      properties:
        name:
          type: string
        path:
          type: string
          description: Absolute server-side archive path.
        size:
          type: integer
          format: int64
          minimum: 0
        createdAt:
          type: string
          format: date-time
        encrypted:
          type: boolean
    BackupCreateRequest:
      type: object
      additionalProperties: false
      properties:
        prune:
          type: boolean
          default: false
        daily:
          type: integer
          minimum: 0
          maximum: 365
          default: 7
        weekly:
          type: integer
          minimum: 0
          maximum: 104
          default: 4
        monthly:
          type: integer
          minimum: 0
          maximum: 120
          default: 12
...
    BackupArchiveFile:
      type: object
      required: [name, size, sha256]
      properties:
        name:
          type: string
          enum: [state.json, state.key, veil.db]
        size:
          type: integer
          format: int64
          minimum: 0
        sha256:
          type: string
          pattern: '^[0-9a-f]{64}$'
    BackupVerificationReport:
      type: object
      required:
        - formatVersion
        - encryptionVersion
        - encrypted
        - legacy
        - stateSchemaVersion
        - desiredRevision
        - files
      properties:
        formatVersion:
          type: integer
          minimum: 0
          description: Zero for legacy archives without a manifest.
        encryptionVersion:
          type: integer
          enum: [0, 1, 2, 3]
          description: Zero for unencrypted archives.
        encrypted:
          type: boolean
        legacy:
          type: boolean
        createdAt:
          type: string
          format: date-time
        veilVersion:
          type: string
        stateSchemaVersion:
          type: integer
          minimum: 1
        desiredRevision:
          type: integer
          format: int64
          minimum: 0
          description: Desired configuration revision captured with the state and SQLite snapshot.
        files:
          type: array
          minItems: 2
          maxItems: 3
          items:
            $ref: "#/components/schemas/BackupArchiveFile"
    BackupPruneResult:
      type: object
      required: [kept, deleted, dryRun]
      properties:
        kept:
          type: array
          items:
            type: string
        deleted:
          type: array
          items:
            type: string
        dryRun:
          type: boolean
    BackupCreateResponse:
      type: object
      required: [archive, verification]
      properties:
        archive:
          $ref: "#/components/schemas/BackupArchive"
        verification:
          $ref: "#/components/schemas/BackupVerificationReport"
        prune:
          $ref: "#/components/schemas/BackupPruneResult"
...